move react_series
