+++ howso/direct/core.py
@@ -727,213 +727,6 @@ class HowsoCore:
             "impute_session": impute_session,
         })
 
-    def batch_react_series(
-        self,
-        trainee_id: str,
-        *,
-        action_features: Optional[Iterable[str]] = None,
-        action_values: Optional[List[List[object]]] = None,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        context_values: Optional[List[List[object]]] = None,
-        context_features: Optional[Iterable[str]] = None,
-        continue_series: Optional[bool] = False,
-        continue_series_features: Optional[Iterable[str]] = None,
-        continue_series_values: Optional[Union[List[object], List[List[object]]]] = None,
-        derived_action_features: Optional[Iterable[str]] = None,
-        derived_context_features: Optional[Iterable[str]] = None,
-        desired_conviction: Optional[float] = None,
-        details: Optional[Dict] = None,
-        exclude_novel_nominals_from_uniqueness_check: bool = False,
-        extra_features: Optional[Iterable[str]] = None,
-        feature_bounds_map: Optional[Dict] = None,
-        final_time_steps: Optional[Union[List[object], List[List[object]]]] = None,
-        generate_new_cases: Literal["always", "attempt", "no"] = "no",
-        init_time_steps: Optional[Union[List[object], List[List[object]]]] = None,
-        initial_features: Optional[Iterable[str]] = None,
-        initial_values: Optional[Union[List[object], List[List[object]]]] = None,
-        input_is_substituted: bool = False,
-        leave_case_out: bool = False,
-        max_series_lengths: Optional[List[int]] = None,
-        new_case_threshold: Literal["max", "min", "most_similar"] = "min",
-        num_series_to_generate: int = 1,
-        ordered_by_specified_features: bool = False,
-        output_new_series_ids: bool = True,
-        preserve_feature_values: Optional[Iterable[str]] = None,
-        series_context_features: Optional[Iterable[str]] = None,
-        series_context_values: Optional[Union[List[object], List[List[object]]]] = None,
-        series_id_tracking: Literal["dynamic", "fixed", "no"] = "fixed",
-        series_stop_maps: Optional[List[Dict[str, Dict]]] = None,
-        substitute_output: bool = True,
-        use_case_weights: bool = False,
-        use_regional_model_residuals: bool = True,
-        weight_feature: Optional[str] = None
-    ) -> Tuple[Dict, int, int]:
-        """
-        React in a series until a series_stop_map condition is met.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        num_series_to_generate : int, optional
-            The number of series to generate.
-        final_time_steps : list of object, optional
-            The time steps at which to end synthesis. Time-series
-            only. Must provide either one for all series, or exactly one per
-            series.
-        init_time_steps : list of object, optional
-            The time steps at which to begin synthesis. Time-series
-            only. Must provide either one for all series, or exactly one per
-            series.
-        initial_features : iterable of str, optional
-            List of features to condition just the first case in a
-            series, overwrites context_features and derived_context_features
-            for that first case. All specified initial features must be in one
-            of: context_features, action_features, derived_context_features or
-            derived_action_features. If provided a value that isn't in one of
-            those lists, it will be ignored.
-        initial_values : list of list of object, optional
-            2d list of values corresponding to the initial_features,
-            used to condition just the first case in each series. Must provide
-            either one for all series, or exactly one per series.
-        series_stop_maps : list of dict of dict, optional
-            A dictionary of feature name to stop conditions. Must provide either
-            one for all series, or exactly one per series.
-        max_series_lengths : list of int, optional
-            maximum size a series is allowed to be.  Default is
-            3 * model_size, a 0 or less is no limit. If forecasting
-            with ``continue_series``, this defines the maximum length of the
-            forecast. Must provide either one for all series, or exactly
-            one per series.
-        continue_series : bool, default False
-            When True will attempt to continue existing series instead of
-            starting new series. If `initial_values` provide series IDs, it
-            will continue those explicitly specified IDs, otherwise it will
-            randomly select series to continue.
-            .. note::
-
-                Terminated series with terminators cannot be continued and
-                will result in null output.
-        continue_series_features : list of str, optional
-            The list of feature names corresponding to the values in each row of
-            `continue_series_values`. This value is ignored if
-            `continue_series_values` is None.
-        continue_series_values : list of list of list of object or list of pandas.DataFrame, default None
-            The set of series data to be forecasted with feature values in the
-            same order defined by `continue_series_values`. The value of
-            `continue_series` will be ignored and treated as true if this value
-            is specified.
-        derived_context_features : iterable of str, optional
-            List of context features whose values should be computed
-            from the entire series in the specified order. Must be
-            different than context_features.
-        derived_action_features : iterable of str, optional
-            List of action features whose values should be computed
-            from the resulting last row in series, in the specified order.
-            Must be a subset of action_features.
-        series_context_features : iterable of str, optional
-            List of context features corresponding to
-            series_context_values, if specified must not overlap with any
-            initial_features or context_features.
-        series_context_values : list of list of list of object or list, optional
-            3d-list of context values, one for each feature for each
-            row for each series. If specified, max_series_lengths are ignored.
-        output_new_series_ids : bool, default True
-            If True, series ids are replaced with unique values on output.
-            If False, will maintain or replace ids with existing trained values,
-            but also allows output of series with duplicate existing ids.
-        series_id_tracking : {"dynamic", "fixed", "no"}, default "fixed"
-            Controls how closely generated series should follow existing series.
-        context_values: list of list of object
-            See parameter ``contexts`` in :meth:`react`.
-        action_features: iterable of str
-            See parameter ``action_features`` in :meth:`react`.
-        action_values: list of list of object
-            See parameter ``actions`` in :meth:`react`.
-        context_features: iterable of str
-            See parameter ``context_features`` in :meth:`react`.
-        input_is_substituted : bool, default False
-            See parameter ``input_is_substituted`` in :meth:`react`.
-        substitute_output : bool
-            See parameter ``substitute_output`` in :meth:`react`.
-        details: dict, optional
-            See parameter ``details`` in :meth:`react`.
-        desired_conviction: float
-            See parameter ``desired_conviction`` in :meth:`react`.
-        exclude_novel_nominals_from_uniqueness_check : bool, default False
-            See parameter ``exclude_novel_nominals_from_uniqueness_check`` in :meth:`react`.
-        weight_feature : str
-            See parameter ``weight_feature`` in :meth:`react`.
-        use_case_weights : bool
-            See parameter ``use_case_weights`` in :meth:`react`.
-        case_indices: iterable of sequence of str, int
-            See parameter ``case_indices`` in :meth:`react`.
-        preserve_feature_values : iterable of str
-            See parameter ``preserve_feature_values`` in :meth:`react`.
-        new_case_threshold : str
-            See parameter ``new_case_threshold`` in :meth:`react`.
-        leave_case_out : bool
-            See parameter ``leave_case_out`` in :meth:`react`.
-        use_regional_model_residuals : bool
-            See parameter ``use_regional_model_residuals`` in :meth:`react`.
-        feature_bounds_map: dict of dict
-            See parameter ``feature_bounds_map`` in :meth:`react`.
-        generate_new_cases : {"always", "attempt", "no"}
-            See parameter ``generate_new_cases`` in :meth:`react`.
-        ordered_by_specified_features : bool
-            See parameter ``ordered_by_specified_features`` in :meth:`react`.
-
-        Returns
-        -------
-        dict
-            A dictionary with keys `action_features` and `series`. Where
-            `series` is a 2d list of values (rows of data per series), and
-            `action_features` is the list of all action features
-            (specified and derived).
-        int
-            The request payload size.
-        int
-            The result payload size.
-        """
-        return self._execute_sized(trainee_id, "batch_react_series", {
-            "context_features": context_features,
-            "context_values": context_values,
-            "action_features": action_features,
-            "action_values": action_values,
-            "final_time_steps": final_time_steps,
-            "init_time_steps": init_time_steps,
-            "initial_features": initial_features,
-            "initial_values": initial_values,
-            "series_stop_maps": series_stop_maps,
-            "max_series_lengths": max_series_lengths,
-            "continue_series": continue_series,
-            "continue_series_features": continue_series_features,
-            "continue_series_values": continue_series_values,
-            "derived_context_features": derived_context_features,
-            "derived_action_features": derived_action_features,
-            "series_context_features": series_context_features,
-            "series_context_values": series_context_values,
-            "series_id_tracking": series_id_tracking,
-            "output_new_series_ids": output_new_series_ids,
-            "details": details,
-            "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
-            "extra_features": extra_features,
-            "case_indices": case_indices,
-            "input_is_substituted": input_is_substituted,
-            "substitute_output": substitute_output,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-            "leave_case_out": leave_case_out,
-            "num_series_to_generate": num_series_to_generate,
-            "preserve_feature_values": preserve_feature_values,
-            "new_case_threshold": new_case_threshold,
-            "use_regional_model_residuals": use_regional_model_residuals,
-            "desired_conviction": desired_conviction,
-            "feature_bounds_map": feature_bounds_map,
-            "generate_new_cases": generate_new_cases,
-            "ordered_by_specified_features": ordered_by_specified_features,
-        })
-
     def get_session_indices(self, trainee_id: str, session: str) -> List[int]:
         """
         Get list of all session indices for a specified session.
